refactor(narratives): address review — filename casing + comment cleanup
Per @nick-nlb's review on PR 2:
- Rename to the app's lowercase category-first convention: App.tsx -> app.tsx,
  Icons.tsx -> icons.tsx, reasoning_block.tsx -> block_reasoning.tsx (with
  imports updated). icons.tsx also re-syncs casing with custom-data-commons.
- Drop change-narrative comments that describe what the code used to do
  (phase-loader removal note, the skeleton 'previously only checked mcp' note).
- metrics_page: the fileoverview described dangerouslySetInnerHTML, which is no
  longer used; point to the <DataCommonsComponent> wrapper instead.
