+++ narratives/src/components/block_reasoning.tsx
