+++ narratives/src/main.tsx
@@ -4,7 +4,7 @@
 
 import {StrictMode} from 'react';
 import {createRoot} from 'react-dom/client';
-import { App } from './App.tsx';
+import { App } from './app.tsx';
 import './index.css';
 
 createRoot(document.getElementById('root')!).render(
